+++ bin/metrics-processes-threads-count.rb
@@ -58,13 +58,6 @@ class ProcessesThreadsCount < Sensu::Plugin::Metric::CLI::Graphite
          boolean: true,
          default: false
 
-  PROCTABLE_MSG = 'sys-proctable version newer than 0.9.5 is required for counting threads with -t or --threads'.freeze
-
-  # Exit with an unknown if sys-proctable is not high enough to support counting threads.
-  def check_proctable_version
-    Gem.loaded_specs['sys-proctable'].version > Gem::Version.create('0.9.5')
-  end
-
   # Takes a value to be tested as an integer. If a new Integer instance cannot be created from it, return 1.
   # See the comments on get_process_threads() for why 1 is returned.
   def test_int(i)
@@ -112,9 +105,6 @@ class ProcessesThreadsCount < Sensu::Plugin::Metric::CLI::Graphite
 
   # Main function
   def run
-    if config[:threads] || config[:statuses]
-      unknown PROCTABLE_MSG unless check_proctable_version
-    end
     ps_table = Sys::ProcTable.ps
     processes = ps_table.length
     threads = count_threads(ps_table) if config[:threads]